♻️ refactor : 전체 호출 위치 변경  #158
