+++ src/components/editor/elements/images/element-image-canvas.tsx
@@ -1,17 +1,16 @@
-/* eslint-disable no-unused-vars */
 'use client';
 import React, { forwardRef } from 'react';
 import { Image as KonvaImage } from 'react-konva';
 import { useImage } from 'react-konva-utils';
 import Konva from 'konva';
-import { ImageElement } from '@/store/editor.store';
+import { ImageElement } from '@/types/editor.type';
 
 interface UnsplashImageElementProps {
   element: ImageElement;
-  onDragEnd: (id: string, node: Konva.Node) => void;
-  onDragMove: (node: Konva.Node) => void;
-  onTransformEnd: (id: string, e: Konva.KonvaEventObject<Event>) => void;
-  onSelect: (id: string, node: Konva.Node) => void;
+  onDragEnd: (_id: string, _node: Konva.Node) => void;
+  onDragMove: (_node: Konva.Node) => void;
+  onTransformEnd: (_id: string, _e: Konva.KonvaEventObject<Event>) => void;
+  onSelect: (_id: string, _node: Konva.Node) => void;
 }
 
 const UnsplashImageElement = forwardRef<Konva.Image, UnsplashImageElementProps>(
