+++ src/components/editor/elements/qr-social/element-qr-canvas.tsx
@@ -1,17 +1,16 @@
-/* eslint-disable no-unused-vars */
 'use client';
 import React, { forwardRef, useEffect } from 'react';
 import { Image as KonvaImage } from 'react-konva';
 import Konva from 'konva';
-import { QrElement } from '@/store/editor.store';
 import { useImage } from 'react-konva-utils';
+import { QrElement } from '@/types/editor.type';
 
 interface QrCanvasElementProps {
   element: QrElement;
-  onDragEnd: (id: string, node: Konva.Node) => void;
-  onDragMove: (node: Konva.Node) => void;
-  onTransformEnd: (id: string, e: Konva.KonvaEventObject<Event>) => void;
-  onSelect: (id: string, node: Konva.Node) => void;
+  onDragEnd: (_id: string, _node: Konva.Node) => void;
+  onDragMove: (_node: Konva.Node) => void;
+  onTransformEnd: (_id: string, _e: Konva.KonvaEventObject<Event>) => void;
+  onSelect: (_id: string, _node: Konva.Node) => void;
 }
 
 const QrCanvasElement = forwardRef<Konva.Image, QrCanvasElementProps>(
