+++ src/components/editor/elements/qr-social/element-social-canvas.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { sideBarStore } from '@/store/editor.sidebar.store';
-import { SocialElement } from '@/store/editor.store';
+import { SocialElement } from '@/types/editor.type';
 import Konva from 'konva';
 import { forwardRef, useEffect } from 'react';
 import { Image as KonvaImage } from 'react-konva';
@@ -9,10 +9,10 @@ import { Html as KonvaHtml, useImage } from 'react-konva-utils';
 
 interface SocialCanvasElementProps {
   element: SocialElement;
-  onDragEnd: (id: string, node: Konva.Node) => void;
-  onDragMove: (node: Konva.Node) => void;
-  onTransformEnd: (id: string, e: Konva.KonvaEventObject<Event>) => void;
-  onSelect: (id: string, node: Konva.Node) => void;
+  onDragEnd: (_id: string, _node: Konva.Node) => void;
+  onDragMove: (_node: Konva.Node) => void;
+  onTransformEnd: (_id: string, _e: Konva.KonvaEventObject<Event>) => void;
+  onSelect: (_id: string, _node: Konva.Node) => void;
 }
 
 const SocialCanvasElement = forwardRef<Konva.Image, SocialCanvasElementProps>(
